long rests to safe rests, potion buff, simplify encounter math, xp variant, minion simplified, no wound from crit to 0

+++ src/pages/Druid/index.tsx
@@ -288,7 +288,7 @@ function Druid() {
               associated damage type, you deal additional damage equal to your
               tier. You can change your realm attunement after a{" "}
               <Link className="internal-link" to="/rules/adventuring#resting">
-                long rest
+                short rest
               </Link>
               .
             </p>
